pyproject.toml
[project]
name = "chatbot-advanced"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.1",
    "langchain>=1.2.17",
    "langchain-core>=1.3.3",
    "langchain-ollama>=1.1.0",
    "langgraph>=1.1.10",
    "langgraph-checkpoint-postgres>=3.0.5",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "psycopg-pool>=3.3.1",
    "psycopg[binary]>=3.3.4",
    "python-dotenv>=1.2.2",
    "streamlit>=1.57.0",
    "uvicorn>=0.46.0",
]
